api/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg[binary,pool]>=3.1
celery>=5.3
redis>=5.0
gunicorn>=22.0
django-cors-headers>=4.3.1

# optional extras you already use
boto3>=1.34
opensearch-py>=2.4
pyarrow>=15.0
duckdb>=0.10
